NEW implement celery tasks for newsletters
