+++ accounts/backends.py
@@ -14,7 +14,7 @@ from .forms import CustomUserRegistrationForm
 
 from organizations.backends.defaults import InvitationBackend
 
-from newsletter.views import register_email
+from newsletter.tasks import register_email
 
 class CustomInvitationsBackend(InvitationBackend):
     """
@@ -87,7 +87,7 @@ class CustomInvitationsBackend(InvitationBackend):
         if form.is_valid():
             form.instance.is_active = True
             if form.cleaned_data['newsletter']: 
-                register_email(form.cleaned_data['email'])
+                register_email.delay(form.cleaned_data['email'])
             user = form.save()
             user.set_password(form.cleaned_data["password1"])
             user.save()
